Restore root span context after span (#24)
* Restore root span context after span

beforeStep replaces the traceCtx.Context when it has created the child
span. This will link spans created with tracer.SpanFromContext on
traceCtx, after a step has been executed, to the step span, instead of
the root span.

So a span created after all steps are executed (e.g. in teardown code)
will be linked to the last span of the test case.

* Use StartSpanFromContext in beforeStep

* Assign to correct variable